refact: 리뷰 반영 - LottosResult에 결과를 반영하는 로직을 생성자로 이동
그에 따른 LottosMoneyService의 함수를 제거하고 Checker 라는 이름으로 변경
