+++ src/main/java/lotto/LottoMoneyChecker.java
@@ -1,0 +1,16 @@
+package lotto;
+
+import lotto.config.LottoGameConfig;
+
+public class LottoMoneyChecker {
+    public int calculatePurchasingCount(int money) {
+        validateMoney(money);
+        return money / LottoGameConfig.PURCHASE_MONEY;
+    }
+
+    private void validateMoney(int money) {
+        if (money < 0) {
+            throw new IllegalArgumentException("money는 음수일 수 없습니다.");
+        }
+    }
+}
